% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smalltableobject.R
\docType{class}
\name{smalltableobject}
\alias{smalltableobject}
\title{Access (small) remote sql-tables just like data frames}
\format{\code{\link{R6Class}} object with class 'smalltableobject'.}
\usage{
sto1 <- smalltableobject$new(...) # Named sql-table on db-server.
sto1[1:10, 1]
sto1[1, 1] <- 99 # Trigger write to db.
sto1[, ] <- sto1[1:10, ] # Must use [,]
}
\value{
Object of \code{\link{R6Class}} abstracting a database table.
}
\description{
Use object to query database table just like a data frame.
E.g tab[1, 1]    or   tab[2, 2] <- 'a'
}
\section{Fields}{

\describe{
\item{\code{get_host}}{Name of host}

\item{\code{get_dbtype}}{Type of DB. Now only "sqlite" allowed.}

\item{\code{get_db}}{Database part of table schema.}

\item{\code{get_tablename}}{Table part of table schema.}
}}

\section{Methods}{

\describe{
  \item{Documentation}{For full source code go to https://github.com/dan070/smalltableobject}
  \item{\code{subset_read}}{Used only with overloaded  operator. Public by neccessity.}
  \item{\code{subset_write}}{Used only with overloaded sS operator. Public by neccessity.}
  \item{\code{print}}{Prints some helpful internal info about object.}
}
}

\examples{
tf <- tempfile()
cn <- RSQLite::dbConnect(RSQLite::SQLite(), tf)
DBI::dbWriteTable(cn, "mytab", cars)
DBI::dbDisconnect(cn)
sto1 <- smalltableobject$new(dbtype = "sqlite", host = tf, tablename = "mytab")
class(sto1)
sto1[1, 1]
}
\keyword{data}
