% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overloaded_operators.R
\name{[.smalltableobject}
\alias{[.smalltableobject}
\title{Subsetting operator for \code{smalltableobject}}
\usage{
\method{[}{smalltableobject}(o, x, y)
}
\arguments{
\item{o}{Object this operates on.}

\item{x}{First argument in []. Missing is valid.}

\item{y}{Second argument in []. Missing is valid.}
}
\value{
Subsetted \code{smalltableobject} based on x and y.
}
\description{
S3-method to enable square brackets on this object.
}
\examples{
sto[1, ] # Show first row of table.
}
