% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overloaded_operators.R
\name{[<-.smalltableobject}
\alias{[<-.smalltableobject}
\title{@rdname add {Assignment operator for \code{smalltableobject}}}
\usage{
\method{[}{smalltableobject}(o, x, y) <- value
}
\arguments{
\item{o}{Object to operate on.}

\item{x}{First argument  \code{[x, ]}. Missing is valid.}

\item{y}{Second argument  \code{[, y]}. Missing is valid.}

\item{value}{Object or value sent in from \code{<-}}
}
\value{
Updated \code{smalltableobject} .
}
\description{
S3-method to enable square bracket assignment on this object.
}
\examples{
no examples

}
