% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smalltableobject.R
\docType{class}
\name{smalltableobject}
\alias{smalltableobject}
\title{Access (small) remote sql-tables just like data frames}
\format{\code{R6Class} object with S3 class 'smalltableobject'.}
\usage{
sto1 <- smalltableobject$new(...) # Named sql-table on db-server.
sto1[1:10, 1]
sto1[1, 1] <- 99 # Trigger a write to db.
sto1[, ] <- sto1[1:10, ] # NB: must use [,]
}
\value{
Database table wrapped in an R6 object.
}
\description{
Use object to query database table just like a data frame, using
...[x, y]  notation.
}
\section{Fields}{

\describe{
\item{\code{get_host}}{Name of host}

\item{\code{get_dbtype}}{Type of DB. Only "sqlite" allowed, currently.}

\item{\code{get_db}}{Database part of table schema.}

\item{\code{get_tablename}}{Table part of table schema.}
}}

\section{Methods}{

\describe{
  \item{Documentation}{For full source code go to \href{https://github.com/dan070/smalltableobject}{Github repo} }
  \item{\code{subset_read}}{Don't call directly. Used with overloaded operator. Public by neccessity.}
  \item{\code{subset_write}}{Don't call directly. Used with overloaded operator. Public by neccessity.}
  \item{\code{print}}{Prints some helpful internal info about object. Also default method of object.}
}
}

\examples{
tf <- tempfile()
cn <- RSQLite::dbConnect(RSQLite::SQLite(), tf)
DBI::dbWriteTable(cn, "mytab", cars)
DBI::dbDisconnect(cn)
sto1 <- smalltableobject$new(dbtype = "sqlite", host = tf, tablename = "mytab")
class(sto1)
sto1[1, 1]
}
\keyword{data}
